pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "djangocms-rest"
dynamic = ["version"]
description = "API endpoints for django CMS"
readme = "README.md"
license = {text = "BSD-3-Clause" }
requires-python = ">=3.9"
authors = [
    { name = "Fabian Braun", email = "fsbraun@gmx.de" },
]
maintainers = [
    { name = "Django CMS Association and contributors", email = "info@django-cms.org" },
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django CMS",
    "Framework :: Django CMS :: 3.11",
    "Framework :: Django CMS :: 4.0",
    "Framework :: Django CMS :: 4.1",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "django-cms>=4.1.1",
    "djangorestframework",
]

[project.urls]
Homepage = "https://github.com/fsbraun/djangocms-rest"

[tool.hatch.version]
path = "djangocms_rest/__init__.py"

[tool.hatch.build]
include = [
    "djangocms_rest",
 ]
artifacts = [
    "djangocms_rest/static/**",
]
